Create list, info page and location page
In order to view more information about a list item, we made the list
items interactive and added an information carousel view for the user to
see detailed information about the list item. We also linked a location
view page as a place holder for a future map. Note: Each page has
navigation buttons for ease of access to all parts of the site.