Fix listview list highlighting on select
I fixed the problem where when we selected a list item and sorted that
store to get the selected vendor's information in order to display the
vendor's products, then went back to the list page the previously selected
item was not highlitghted. This occured, because when we sorted by
selected vendor the list item ids were reset, so the program didn't know
which item to highlight when we navigated back to the list page. In the
future this will be solved with an api call, for now I just made another
vendor store that we could sort to get the specific vendor data needed for
the detail page.